chore: organize story files of organisms
