+++ src/stories/designSystem/organisms/Footer.stories.tsx
@@ -1,14 +1,18 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Intro from "./Intro";
+import { Footer } from "@/ui";
 
-const meta: Meta<typeof Intro> = {
-  title: "organisms/Intro",
-  component: Intro,
+const meta: Meta<typeof Footer> = {
+  title: "design system/organisms/Footer",
+  component: Footer,
   tags: ["autodocs"],
   parameters: {
     controls: { expanded: true },
+    docs: {
+      description: {
+        component: "Footer section",
+      },
+    },
   },
-  argTypes: {},
 };
 export default meta;
