+++ src/stories/designSystem/organisms/Hero.stories.tsx
@@ -1,14 +1,18 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Hero from "./Hero";
+import { Hero } from "@/ui";
 
 const meta: Meta<typeof Hero> = {
-  title: "organisms/Hero",
+  title: "design system/organisms/Hero",
   component: Hero,
   tags: ["autodocs"],
   parameters: {
     controls: { expanded: true },
+    docs: {
+      description: {
+        component: "Hero section",
+      },
+    },
   },
-  argTypes: {},
 };
 export default meta;
